requirements-attachments.txt
# Additional requirements for attachment processing
pdfplumber>=0.10.0  # PDF text extraction
pytesseract>=0.3.10  # OCR for images
Pillow>=10.0.0  # Image processing
python-docx>=1.1.0  # Word document text extraction
